int -> double

+++ app/src/main/java/com/example/dartjonny/dart_jonny/model/Player.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class Player(
     val playerName: String,
-    val points: Double = 0.0,
+    val points: Int = 0,
     val color: Int,
     @PrimaryKey(autoGenerate = true) val id: Int? = null
 )
